web/requirements.txt
Flask
flask-restful
pymongo
bcrypt
numpy
tensorflow
requests
keras
pillow